requirements.txt
coreapi
mod_wsgi
Django
django-allauth
django-rest-auth[with_social]
djangorestframework
git+https://github.com/vincent-noel/GINsim-python.git
git+https://github.com/vincent-noel/pyMaBoSS.git
git+https://github.com/vincent-noel/MaBoSS-Client.git
mysqlclient
PyYAML